--- mesofield/data/proc/__init__.py
@@ -1,1 +1,0 @@
-from .load import file_hierarchy, ExperimentData
