--- mesofield/io/__init__.py
@@ -1,3 +1,0 @@
-from ..data.writer import CustomWriter, CV2Writer
-from .h5db import H5Database
-from . import sessiondb
