--- mesofield/io/devices/__init__.py
@@ -1,18 +1,0 @@
-from .base import BaseDevice, BaseDataProducer, BaseSerialDevice
-from .daq import Nidaq
-from .cameras import MMCamera, OpenCVCamera
-from .encoder import SerialWorker
-from .treadmill import EncoderSerialInterface
-from .psychopy_device import PsychoPyDevice
-
-__all__ = [
-    "BaseDevice",
-    "BaseDataProducer",
-    "BaseSerialDevice",
-    "Nidaq",
-    "MMCamera",
-    "OpenCVCamera",
-    "SerialWorker",
-    "EncoderSerialInterface",
-    "PsychoPyDevice",
-]
