--- mesofield/protocols/__init__.py
@@ -1,240 +1,0 @@
-"""
-Protocol definitions for hardware instruments and data management.
-
-This module defines the core interfaces that standardize behavior across
-the mesofield project, allowing for interoperability between different
-hardware instruments, data producers, and data consumers.
-
-Protocol Implementation Notes
-----------------------------
-When implementing these protocols, there are two approaches:
-
-1. Direct inheritance (for regular classes without metaclass conflicts):
-   ```python
-   class MySensor(DataAcquisitionDevice):
-       def __init__(self):
-           self._init_logger()  # Initialize logging
-           # Implement required methods and attributes
-   ```
-
-2. Duck typing (for classes with existing inheritance or metaclass conflicts, e.g., QThread):
-   ```python
-   class MyQThreadSensor(QThread):  # Cannot inherit from Protocol due to metaclass conflict
-       def __init__(self):
-           super().__init__()
-           self._init_logger()  # Initialize logging manually
-           # Implement all required methods and attributes
-           self.device_type = "sensor"
-           self.device_id = "my_sensor"
-           # etc.
-   ```
-
-The second approach is necessary for Qt classes (QObject, QThread, QWidget) or 
-any class that already uses a metaclass. Protocol checking uses duck typing 
-internally, so both approaches will work with our system.
-"""
-
-from typing import Dict, List, Any, Optional, Protocol, TypeVar, Generic, runtime_checkable
-
-from typing import TYPE_CHECKING
-if TYPE_CHECKING:
-    from mesofield.hardware import HardwareManager
-
-T = TypeVar('T')
-
-# These are the Protocol definitions - they are useful for static type checking
-# and documentation, but should not be used for inheritance with classes that
-# already have a metaclass (like QThread)
-
-
-# Define configuration interface
-class Configurator(Protocol):
-    """Protocol defining the interface for configuration providers."""
-    
-    hardware: "HardwareManager"
-    
-    def get(self, key: str, default: Any = None) -> Any:
-        """Retrieve a configuration value for the given key."""
-        ...
-    
-    def set(self, key: str, value: Any) -> None:
-        """Set a configuration value for the given key."""
-        ...
-    
-    def has(self, key: str) -> bool:
-        """Check if the configuration contains the given key."""
-        ...
-    
-    def keys(self) -> List[str]:
-        """Get all configuration keys."""
-        ...
-    
-    def items(self) -> Dict[str, Any]:
-        """Get all configuration key-value pairs."""
-        ...
-
-class Procedure(Protocol):
-    """Protocol defining the standard interface for experiment procedures."""
-    
-    protocol: str
-    experimenter: str
-    config: Configurator
-    data_dir: str
-    
-    def initialize_hardware(self) -> None:
-        """Setup the experiment procedure.
-        
-        """
-        ...
-    
-    def setup_configuration(self, json_config: Optional[str]) -> None:
-        """Set up the configuration for the experiment procedure.
-        
-        Args:
-            json_config: Path to a JSON configuration file (.json)
-        """
-        ...    
-        
-    def run(self) -> None:
-        """Run the experiment procedure."""
-        ...
-        
-    def save_data(self) -> None:
-        """Save data from the experiment."""
-        ...
-        
-    def cleanup(self) -> None:
-        """Clean up after the experiment procedure."""
-        ...
-        
-@runtime_checkable
-class HardwareDevice(Protocol):
-    """Protocol defining the standard interface for all hardware devices.
-
-    Lifecycle: ``initialize`` -> ``arm`` -> ``start`` -> ``stop`` -> ``shutdown``.
-    Every device exposes ``self.signals`` (a :class:`mesofield.signals.DeviceSignals`)
-    carrying ``started``, ``finished``, and ``data`` emitters.
-    """
-
-    device_type: str
-    device_id: str
-    signals: Any  # DeviceSignals; typed as Any to avoid circular import
-
-    def initialize(self) -> bool:
-        """One-time setup (open ports, load configs).  Idempotent."""
-        ...
-
-    def arm(self, config: Any) -> None:
-        """Per-run preparation (writers, output paths, sequence build).
-
-        Called by ``HardwareManager.arm_all`` immediately before
-        ``start_all``.  Devices without per-run prep may no-op.
-        """
-        ...
-
-    def stop(self):
-        """Stop the device after a run."""
-        ...
-
-    def shutdown(self) -> None:
-        """Close and clean up resources."""
-        ...
-
-    def status(self) -> Dict[str, Any]:
-        """Get the current status of the device."""
-        ...
-
-    @property
-    def metadata(self) -> Dict[str, Any]:
-        """Return metadata about the hardware."""
-        ...
-
-
-
-@runtime_checkable
-class DataProducer(HardwareDevice, Protocol):
-    """Protocol for hardware that produces data streamed to the DataQueue."""
-
-    sampling_rate: float  # in Hz
-    data_type: str
-    file_type: str
-    bids_type: Optional[str] = None
-    is_active: bool
-    output_path: str
-    metadata_path: Optional[str] = None
-
-    def start(self) -> bool:
-        """Start data acquisition.  Should emit ``signals.started``."""
-        ...
-
-    def stop(self) -> bool:
-        """Stop data acquisition.  Should emit ``signals.finished``."""
-        ...
-
-    def save_data(self, path: Optional[str] = None):
-        """Persist the captured data."""
-        ...
-
-    def get_data(self) -> Optional[Any]:
-        """Return the latest data."""
-        ...
-
-
-@runtime_checkable
-class StimulusDevice(HardwareDevice, Protocol):
-    """Protocol for stimulus-presentation devices (e.g. PsychoPy).
-
-    Like :class:`HardwareDevice` but explicitly *not* a data producer:
-    consumers should not expect ``data`` signal emissions and should not
-    call ``save_data``/``get_data``.
-    """
-
-    def start(self) -> bool:
-        """Begin stimulus presentation."""
-        ...
-    
-
-
-
-@runtime_checkable
-class DataConsumer(Protocol):
-    """Protocol defining the interface for data-consuming components."""
-    
-    @property
-    def name(self) -> str:
-        """Return the name of the data consumer."""
-        ...
-    
-    @property
-    def get_supported_data_types(self) -> List[str]:
-        """Return the types of data this consumer can process."""
-        ...
-    
-    def process_data(self, data: Any, metadata: Dict[str, Any]) -> bool:
-        """Process data with metadata.
-        
-        Args:
-            data: The data to process.
-            metadata: Metadata about the data, including source, timestamp, etc.
-            
-        Returns:
-            bool: True if data was processed successfully, False otherwise.
-        """
-        ...
-        
-        
-# Helper functions for protocol checking
-
-def is_hardware_device(obj) -> bool:
-    """Check if an object implements the HardwareDevice interface."""
-    required_attrs = ['device_id', 'device_type', 'config', 'initialize', 
-                     'start', 'stop', 'close', 'get_status']
-    return all(hasattr(obj, attr) for attr in required_attrs)
-
-def is_data_acquisition_device(obj) -> bool:
-    """Check if an object implements the DataAcquisitionDevice interface."""
-    if not is_hardware_device(obj):
-        return False
-    return hasattr(obj, 'data_rate') and hasattr(obj, 'get_data')
-
-
