+++ mesofield/scaffold/experiment.py
@@ -219,7 +219,7 @@ def _procedure_py(protocol: str) -> str:
         '\n'
         'from mesofield import DeviceRegistry\n'
         'from mesofield.base import Procedure\n'
-        'from mesofield.examples.mock_encoder import MockEncoderDevice\n'
+        'from mesofield.devices.mocks import MockEncoderDevice\n'
         '\n'
         '# Register any custom devices you have added under devices/ here.\n'
         '# Built-in device types (camera, wheel, encoder, psychopy, nidaq)\n'
@@ -318,7 +318,7 @@ def _thermal_example_py() -> str:
         'import pandas as pd\n'
         '\n'
         'from mesofield import DeviceRegistry\n'
-        'from mesofield.io.devices.base import BaseSerialDevice\n'
+        'from mesofield.devices.base import BaseSerialDevice\n'
         'from mesofield.datakit.sources.register import SourceContext, TimeseriesSource\n'
         '\n'
         '\n'
