+++ tests/test_opencv_camera.py
@@ -1,4 +1,4 @@
-"""Integration test for :class:`mesofield.io.devices.OpenCVCamera`.
+"""Integration test for :class:`mesofield.devices.OpenCVCamera`.
 
 Records a few seconds from the local OpenCV camera, verifies that:
   * Protocol attributes are present (HardwareDevice / DataProducer).
@@ -20,7 +20,7 @@ pytest.importorskip("cv2")
 from PyQt6.QtCore import QCoreApplication  # noqa: E402
 
 from mesofield.data.manager import DataManager  # noqa: E402
-from mesofield.io.devices.cameras import OpenCVCamera, _resolve_cv_backend  # noqa: E402
+from mesofield.devices.cameras import OpenCVCamera, _resolve_cv_backend  # noqa: E402
 from mesofield.protocols import DataProducer, HardwareDevice  # noqa: E402
